refactor: move active filter chips to separate component
